refactor: update Qubic import paths to use the new core module
