+++ examples/chain/getBlockHeight.ts
@@ -1,4 +1,4 @@
-import Qubic from "../../src/qubic";
+import Qubic from "../../src/core";
 
 async function getBlockHeight() {
   // Initialize the Qubic instance with the provider URL
