+++ examples/chain/getComputors.ts
@@ -1,4 +1,4 @@
-import Qubic from "../../src/qubic";
+import Qubic from "../../src/core";
 
 async function getComputors() {
   // Initialize the Qubic instance with the provider URL
