+++ examples/identity/getBalanceByIdentity.ts
@@ -1,4 +1,4 @@
-import Qubic from "../../src/qubic";
+import Qubic from "../../src/core";
 
 async function getBalanceByIdentity() {
   try {
