+++ examples/transaction/getApprovedTransactions.ts
@@ -1,4 +1,4 @@
-import Qubic from "../../src/qubic";
+import Qubic from "../../src/core";
 
 async function getApprovedTransactions() {
   // Initialize the Qubic instance with the provider URL
